TST: Migrated some tests from test_atomgroup to universe
